Fix impossible to choose a path in the forest maze
Fixes #187
References #38

+++ include/Entities/Actor/BlinkState.hpp
@@ -8,7 +8,7 @@ class BlinkState {
 public:
   explicit BlinkState(Costume &costume);
 
-  void setRate(double min, double max);
+  void setRate(float min, float max);
   void update(sf::Time elapsed);
 
 private:
